test(bot): consolidate rapid helpers and clarify contract tests
- introduce rapid_helpers_test.go with a single
  genSupportedCurrency, replacing three near-duplicates across
  parser, csv_generator, and currency_conversion tests so
  future SupportedCurrencies changes propagate uniformly
- cache the sorted code slice via sync.OnceValue so
  rapid.Check loops stop rebuilding it per iteration
- document TestIsStopWordFixedSet as a deliberate pin on the
  current stop-word set and TestMatchCategorySubstringFinds as
  a non-nil containment contract that does not constrain which
  candidate wins, so future behavior tweaks are easier to
  reason about
